refactor:Make the random list and favorite list into a reusable component.
