--- src/components/SmoothScroll.tsx
@@ -1,15 +1,0 @@
-// import{ useRef } from 'react';
-// import ExpandCircleDownRoundedIcon from '@mui/icons-material/ExpandCircleDownRounded';
-// const SmoothScroll = () => {
-//     const ref=useRef(null);
-//     const handleClick=()=>{
-//         ref.current?.scrollIntoView({behavior:'smooth'});
-//     };
-//   return (
-//     <button onClick={handleClick} className="bottom-0 left-1/2  text-6xl  bg-white absolute">
-//       <ExpandCircleDownRoundedIcon />
-//     </button>
-//   );
-// };
-
-// export default SmoothScroll;
